Feat: Integration-config update
